# openapi/foodize.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/promos": {
      "post": {
        "tags": [
          "Promos"
        ],
        "summary": "Create Promo",
        "operationId": "create_promo_api_v1_promos_post",
        "security": [
          {
            "OAuth2PasswordBearerWithCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_PromoResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Promos"
        ],
        "summary": "List Promos",
        "operationId": "list_promos_api_v1_promos_get",
        "security": [
          {
            "OAuth2PasswordBearerWithCookie": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessListResponse_PromoResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/promos/{code}": {
      "delete": {
        "tags": [
          "Promos"
        ],
        "summary": "Deactivate Promo",
        "operationId": "deactivate_promo_api_v1_promos__code__delete",
        "security": [
          {
            "OAuth2PasswordBearerWithCookie": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/promos/validate": {
      "post": {
        "tags": [
          "Promos"
        ],
        "summary": "Validate Promo",
        "operationId": "validate_promo_api_v1_promos_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_PromoValidateResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearerWithCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Meta": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "Meta"
      },
...
      "Pagination": {
        "properties": {
          "current_page": {
            "type": "integer",
            "title": "Current Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          },
          "previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous"
          }
        },
        "type": "object",
        "required": [
          "current_page",
          "per_page",
          "total",
          "total_pages",
          "next",
          "previous"
        ],
        "title": "Pagination"
      },
...
      "PromoCreate": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "title": "Code"
          },
          "discount_type": {
            "type": "string",
            "enum": [
              "PERCENT",
              "FIXED"
            ],
            "title": "Discount Type"
          },
          "discount_value": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Discount Value"
          },
          "restaurant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Restaurant Id"
          },
          "max_uses": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "code",
          "discount_type",
          "discount_value",
          "restaurant_id"
        ],
        "title": "PromoCreate"
      },
      "PromoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "discount_type": {
            "type": "string",
            "title": "Discount Type"
          },
          "discount_value": {
            "type": "integer",
            "title": "Discount Value"
          },
          "restaurant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Restaurant Id"
          },
          "max_uses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          },
          "used_count": {
            "type": "integer",
            "title": "Used Count"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "discount_type",
          "discount_value",
          "restaurant_id",
          "max_uses",
          "used_count",
          "expires_at",
          "is_active",
          "created_at"
        ],
        "title": "PromoResponse"
      },
      "PromoValidateRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "restaurant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Restaurant Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "restaurant_id"
        ],
        "title": "PromoValidateRequest"
      },
      "PromoValidateResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "discount_type": {
            "type": "string",
            "title": "Discount Type"
          },
          "discount_value": {
            "type": "integer",
            "title": "Discount Value"
          },
          "discounted_amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discounted Amount"
          }
        },
        "type": "object",
        "required": [
          "code",
          "discount_type",
          "discount_value"
        ],
        "title": "PromoValidateResponse"
      },
...
      "SuccessListResponse_PromoResponse_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PromoResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "SuccessListResponse[PromoResponse]"
      },
...
      "SuccessResponse_PromoResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PromoResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[PromoResponse]"
      },
      "SuccessResponse_PromoValidateResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PromoValidateResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[PromoValidateResponse]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearerWithCookie": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}